final 2

+++ app/Http/Controllers/ColorController.php
@@ -86,29 +86,7 @@ class ColorController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    // public function update(ColorEditRequest $request, Color $color): RedirectResponse
-    // {
-    //     $color->update($request->validated());
-    //     $url = route('colors.show', ['color' => $color]);
-    //     $htmlMessage = "Cor <a href='$url'>#{$color->code}</a> foi alterada com sucesso!";
-    //     return redirect()->route('colors.index')
-    //         ->with('alert-msg', $htmlMessage)
-    //         ->with('alert-type', 'success');
 
-    //     // $formData = $request->validated();
-    //     // $color = DB::transaction(function () use ($formData, $color) {
-    //     //     $color->name = $formData['name'];
-    //     //     //$color->code = $formData['code'];
-    //     //     $color->save();
-    //     //     return $color;
-    //     // });
-    //     // $url = route('colors.show', ['color' => $color]);
-    //     // $htmlMessage = "Cor <a href='$url'>#{$color->code}</a>
-    //     //                 <strong>\"{$color->name}\"</strong> foi alterada com sucesso!";
-    //     //     return redirect()->back()
-    //     //         ->with('alert-msg', $htmlMessage)
-    //     //         ->with('alert-type', 'success');
-    // }
     public function update(ColorEditRequest $request, Color $color): RedirectResponse
     {
         $color->name = $request->input('name');
@@ -125,54 +103,40 @@ class ColorController extends Controller
     /**
      * Remove the specified resource from storage.
      */
+
     public function destroy(Color $color): RedirectResponse
     {
+
         try {
-            $color->delete();
-            $htmlMessage = "Cor #{$color->code}foi apagada com sucesso!";
-            $alertType = 'success';
+            $totalColors = DB::scalar('select count(*) from order_items where color_code = ?', [$color->code]);
+            if ($totalColors == 0) {
+                $color->delete();
+                $htmlMessage = "Cor #{$color->code}
+                                <strong>\"{$color->name}\"</strong> foi apagada com sucesso!";
+                return redirect()->route('colors.index')
+                    ->with('alert-msg', $htmlMessage)
+                    ->with('alert-type', 'success');
+            } else {
+                $url = route('colors.show', ['color' => $color]);
+                $alertType = 'warning';
+                $colorsStr = $totalColors > 0 ?
+                    ($totalColors == 1 ?
+                        "1 Order Item com esta cor" :
+                        "$totalColors Order Items com esta cor") :
+                    "";
+                $htmlMessage = "Cor <a href='$url'>#{$color->code}</a>
+                                <strong>\"{$color->name}\"</strong>
+                                não pode ser apagada porque há $colorsStr!
+                                ";
+            }
         } catch (\Exception $error) {
             $url = route('colors.show', ['color' => $color]);
-            $htmlMessage = "Não foi possível apagar a Cor <a href='$url'>#{$color->code}</a> porque ocorreu um erro!";
+            $htmlMessage = "Não foi possível apagar a cor <a href='$url'>#{$color->code}</a>
+                                <strong>\"{$color->name}\"</strong> porque ocorreu um erro!";
             $alertType = 'danger';
         }
-        return redirect()->route('colors.index')
+        return back()
             ->with('alert-msg', $htmlMessage)
             ->with('alert-type', $alertType);
-
-        //TODO: Utilizar o de baixo quando ja tenhamos criado o order_items
-
-        // try {
-        //     $totalOrderItems = DB::scalar('select count(*) from order_items where color_code = ?', [$color->code]);
-        //     if ($totalOrderItems == 0) {
-        //         $color->delete();
-        //         $htmlMessage = "Cor #{$color->id}
-        //                 <strong>\"{$color->name}\"</strong> foi apagada com sucesso!";
-        //         return redirect()->route('colors.index')
-        //             ->with('alert-msg', $htmlMessage)
-        //             ->with('alert-type', 'success');
-        //     } else {
-        //         $url = route('colors.show', ['color' => $color]);
-        //         $alertType = 'warning';
-        //         $orderItemsStr = $totalOrderItems > 0 ?
-        //             ($totalOrderItems == 1 ?
-        //                 "1 Order Item com esta cor" :
-        //                 "$totalOrderItems Order Items com esta cor") :
-        //             "";
-        //             $htmlMessage = "Cor <a href='$url'>#{$color->id}</a>
-        //                 <strong>\"{$color->name}\"</strong>
-        //                 não pode ser apagada porque há $orderItemsStr!
-        //                 ";
-        //     }
-        // } catch (\Exception $error) {
-        //     $url = route('colors.show', ['color' => $color]);
-        //     $htmlMessage = "Não foi possível apagar a cor <a href='$url'>#{$color->id}</a>
-        //                 <strong>\"{$color->name}\"</strong> porque ocorreu um erro!";
-        //     $alertType = 'danger';
-        // }
-        // return back()
-        //     ->with('alert-msg', $htmlMessage)
-        //     ->with('alert-type', $alertType);
-
     }
 }
